Move validations folder into the Users page
