+++ packages/core/admin/admin/src/pages/ProfilePage/utils/schema.js
@@ -1,5 +1,5 @@
 import * as yup from 'yup';
-import { profileValidation } from '../../../validations/users';
+import { profileValidation } from '../../Users/utils/validations/users';
 
 const schema = yup.object().shape(profileValidation);
